+++ backend/src/listener/contract-listener.ts
@@ -1,41 +1,6 @@
 import { CallData } from "starknet";
 import { EventEmitter } from "stream";
-
-export interface ContractListenerOptions {
-    rpcUrl: string;
-    contractAddress: string;
-    fromBlock: number;
-    chunkSize?: number;
-    abi: any;
-    decodeParameters: string[];
-}
-
-export interface ContractEvent {
-    block_hash: string;
-    block_number: number;
-    data: string[]
-    from_address: string;
-    keys: string[];
-    transaction_hash: string;
-}
-
-interface StarknetResult {
-    calldata: string[];
-    max_fee: string;
-    nonce: string;
-    sender_address: string;
-    signature: string[];
-    transaction_hash: string;
-    type: string;
-    version: string;
-}
-
-export interface StarknetTransaction {
-    jsonrpc: string;
-    id: number;
-    method: string;
-    result: StarknetResult
-}
+import { ContractEvent, ContractListenerOptions, StarknetTransaction } from "../types";
 
 export class ContractListener extends EventEmitter {
     private options: ContractListenerOptions;
